23.1

+++ tests/suite/64_byte_literals.rs
@@ -3,12 +3,12 @@
 // out: 72 101 108 108 111
 
 fn main() {
-    let a: u32 = b'A';
+    let a: u8 = b'A';
     println!("{}", a);
 
-    let h: u32 = b'h';
+    let h: u8 = b'h';
     println!("{}", h);
 
-    let bytes: [u32; 5] = [b'H', b'e', b'l', b'l', b'o'];
+    let bytes: [u8; 5] = [b'H', b'e', b'l', b'l', b'o'];
     println!("{} {} {} {} {}", bytes[0], bytes[1], bytes[2], bytes[3], bytes[4]);
 }
